# protocol/v1/openapi/sealed-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZROtext sealed API v1 (proposed contract)",
    "version": "1.2.0",
    "description": "CONTRACT WITH DEFAULT-OFF IMPLEMENTED SLICES, derived from ZT-009 decisions recorded 2026-09-26. IMPLEMENTED: POST /v1/sealed/messages, POST /v1/sealed/inbound-events, metadata-only message list/status/pre-grant cancellation, and the read-only devices, webhook metadata/delivery metadata and usage resource groups. All are mounted only under SEALED_ADMISSION_ENABLED; the runtime is disabled by default. Admission/upload require exact raw binary envelopes; lifecycle operations never expose content or recipient fields. Existing slice-2 resource handlers and tests define their current projections; the prospective resource schemas here are not evidence that every proposed field or error is implemented. PROPOSAL ONLY: webhook mutations and remaining prospective resource behavior; no server route implements those proposals. Ciphertext retrieval, negotiated grants and outbound dispatch-status fanout remain deferred. The synthetic-alpha plane and stable general-send plane stay separate. There is no JSON/plaintext admission, downgrade, fallback translation, implicit reader or carrier-delivery evidence. All implemented responses are no-store."
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/v1/sealed/messages": {
      "post": {
        "tags": ["sealed-messages"],
        "operationId": "submitSealedOutbound",
        "summary": "Submit one sealed outbound envelope for device dispatch",
        "description": "Proposed contract; no route exists. The request body is exactly one complete sealed outbound envelope (profile kind 01) as raw binary bytes with the exact content type application/vnd.zrotext.sealed.v1. The relay never decrypts: it verifies the owner-pinned manifest chain, the allowed device signing key and scope, and the device signature over the exact unsigned bytes, then compares the protected account/device/line/manifest fields with the authenticated identity and database state. Idempotency is the SHA-256 digest of the unsigned envelope bytes (Q6): an identical replay returns the stored message identity with created=false and performs no new dispatch; no caller-supplied idempotency-key header is accepted. Admission fails closed when the active manifest is expired or was issued too far in the future (Q4), and clock-rollback or restored-high-water conditions require fresh owner-verified enrollment (Q4). The plaintext inside the ciphertext is a strict UTF-8 body of 1-32,768 bytes with no BOM, no NUL, no Unicode normalization and no padding in v1 (Q9); body length, recipient, timing and segment-count leakage is disclosed and accepted. Every response carries Cache-Control: no-store.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "description": "One complete sealed outbound envelope (kind 01) as raw bytes. Bounds: the envelope is at least 426 bytes, at most 34,213 bytes for kind 01, and the request-body limit is the 36,864-byte pre-allocation cap. A JSON body, any other content type, an unknown profile or kind, a set downgrade flag, or trailing bytes is rejected (Q11).",
          "content": {
            "application/vnd.zrotext.sealed.v1": {
              "schema": {
                "$ref": "#/components/schemas/SealedEnvelopeBytes"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted for dispatch, or an identical replay no-op. Acceptance means durable storage of the exact envelope bytes and queueing toward the bound device; it never means carrier submission. Idempotency is on the unsigned envelope digest (Q6): replaying byte-identical unsigned content returns the existing message identity with created=false.",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SealedOutboundAccepted"
                }
              }
            }
          },
          "400": {
            "description": "invalid_request: the bytes are not a bounded, exactly parsed sealed envelope, or verification of the manifest chain, signing key/scope or device signature failed, or a client-allocated identity is malformed. future_manifest: the manifest's issued_at is more than five minutes ahead of trusted server time (Q4); resubmitting unchanged will not help until the manifest is reissued.",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "invalidEnvelope": {
                    "value": { "code": "invalid_request" }
                  },
                  "futureManifest": {
                    "value": { "code": "future_manifest" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized: missing, malformed or unknown bearer credential.",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": { "code": "unauthorized" }
              }
            }
          },
          "403": {
            "description": "forbidden: the credential is not authorized for the tenant, line or device named in the envelope, or the key is revoked or wrong-role. stale_manifest: the active manifest has expired (its exclusive expiry passed, or its lifetime exceeds 24 hours), which blocks sealed outbound fail-closed until a fresh manifest is pinned (Q4). re_enrollment_required: a device-clock rollback or a restored high-water store was detected; sealed submission stays refused until fresh owner-verified enrollment (Q4).",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "forbidden": {
                    "value": { "code": "forbidden" }
                  },
                  "staleManifest": {
                    "value": { "code": "stale_manifest" }
                  },
                  "reEnrollmentRequired": {
                    "value": { "code": "re_enrollment_required" }
                  }
                }
              }
            }
          },
          "409": {
            "description": "idempotency_conflict: a client-allocated identity in the protected record was already stored under a different unsigned envelope digest (Q6); the earlier envelope stays authoritative and this submission creates nothing.",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": { "code": "idempotency_conflict" }
              }
            }
          },
          "415": {
            "description": "unsupported_media_type: the request did not use exactly application/vnd.zrotext.sealed.v1. A JSON body, a multipart body, or any other content type is rejected here and is never parsed as a sealed envelope, translated to the synthetic-alpha route, or retried as plaintext (Q11).",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": { "code": "unsupported_media_type" }
              }
            }
          },
          "429": {
            "description": "rate_limited, queue_full or quota_exceeded: the account or device exceeded its sealed admission budget.",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": { "code": "rate_limited" }
              }
            }
          },
          "503": {
            "description": "unavailable or billing_pending: storage or dispatch is temporarily unavailable; retry later with the same envelope bytes.",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": { "code": "unavailable" }
              }
            }
          }
        }
      },
        "get": {
          "tags": [
            "sealed-messages"
          ],
          "operationId": "listSealedMessages",
          "summary": "listSealedMessages",
          "description": "messages:read. Twenty newest records ordered by (created_at,id), with one optional exclusive cursor resolved inside the same account/device/transport visibility. Foreign/deleted cursor is indistinguishable from absent. Existing sealed queue identity; no new queue, admission verifier, plaintext alternative or dispatch enablement. Mounted only behind SEALED_ADMISSION_ENABLED, disabled by default. Credentials are revalidated under transaction locks; account/device-bound visibility matches current scopes. Metadata only, no envelope or recipient fields. Cache-Control: no-store.",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "Current stored metadata; cancellation is idempotent.",
              "headers": {
                "Cache-Control": {
                  "$ref": "#/components/headers/NoStore"
                }
              },
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SealedMessagePage"
                  }
                }
              }
            },
            "400": {
              "description": "invalid_request: malformed UUID/query or nonempty cancellation body",
              "headers": {
                "Cache-Control": {
                  "$ref": "#/components/headers/NoStore"
                }
              },
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "401": {
              "description": "unauthorized: missing, malformed, expired or revoked bearer",
              "headers": {
                "Cache-Control": {
                  "$ref": "#/components/headers/NoStore"
                }
              },
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "403": {
              "description": "forbidden: wrong scope or authority changed after authentication",
              "headers": {
                "Cache-Control": {
                  "$ref": "#/components/headers/NoStore"
                }
              },
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "404": {
              "description": "not_found: disabled route, absent, foreign, wrong-device or non-sealed identity/cursor",
              "headers": {
                "Cache-Control": {
                  "$ref": "#/components/headers/NoStore"
                }
              },
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "429": {
              "description": "rate_limited: account in-flight budget exhausted",
              "headers": {
                "Cache-Control": {
                  "$ref": "#/components/headers/NoStore"
                }
              },
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "503": {
              "description": "unavailable: database/storage failure",
              "headers": {
                "Cache-Control": {
                  "$ref": "#/components/headers/NoStore"
                }
              },
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          },
          "parameters": [
            {
              "name": "cursor",
              "in": "query",
              "required": false,
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            }
          ]
        }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "The existing customer API-key authentication, not a new auth model: an account API key issued by the owner API-key routes under /v1/auth and sent as 'Authorization: Bearer <key>', exactly as the existing message routes authenticate. Outbound submission requires the existing message-send scope bound to the device named in the envelope's protected record; inbound upload requires a credential authorized for the uploading device, and the protected account, device and line fields must match the authenticated identity and the active line-binding generation. The slice 2 resource operations reuse the existing scope set and add no new scopes: devices list/status requires devices:read, webhook reads require webhooks:read, webhook management (create, replay, enable, disable, rotate) requires webhooks:manage, and the usage query requires billing:read. API keys never unlock content: no scope grants decryption, plaintext access or envelope bytes beyond the exact bytes a caller itself submitted. This HTTP layer only names the tenant: the sealed manifest chain and device signature over the exact unsigned bytes carry the cryptographic authorization (Q2, Q7). Sealed and alpha credentials stay separate (Q11)."
      }
    },
    "headers": {
      "NoStore": {
        "description": "Every sealed response is non-cacheable, mirroring the existing message routes.",
        "schema": {
          "type": "string",
          "const": "no-store"
        }
      }
    },
    "schemas": {
      "SealedEnvelopeBytes": {
        "type": "string",
        "format": "binary",
        "description": "One complete sealed envelope as raw binary bytes: the exact profile bytes including the fixed prefix, protected record, body ciphertext, recipient wraps and signature. Never JSON, never base64-wrapped, never a multipart part. The encrypted plaintext is a strict UTF-8 body of 1-32,768 bytes with no BOM, no NUL, no Unicode normalization and no padding in v1 (Q9)."
      },
      "SealedOutboundAccepted": {
        "type": "object",
        "description": "Acceptance metadata for a sealed outbound envelope. Contains no recipient, no plaintext and no envelope bytes.",
        "required": ["message_id", "created"],
        "properties": {
          "message_id": {
            "type": "string",
            "format": "uuid",
            "description": "The server-allocated message identity for the accepted envelope; identical on digest replays."
          },
          "created": {
            "type": "boolean",
            "description": "false when this exact unsigned envelope digest was already accepted (idempotent replay, Q6)."
          }
        },
        "additionalProperties": false
      },
...
      "Error": {
        "type": "object",
        "description": "The existing error shape used by the message routes: a single machine-readable code and no content echo.",
        "required": ["code"],
        "properties": {
          "code": {
            "type": "string",
            "description": "Lowercase snake_case machine code for the admission or resource failure; the response description defines the taxonomy.",
            "enum": [
            "cancellation_conflict",
              "invalid_request",
              "invalid_webhook_endpoint",
              "future_manifest",
              "stale_event",
              "unauthorized",
              "forbidden",
              "not_found",
              "stale_manifest",
              "re_enrollment_required",
              "idempotency_conflict",
              "event_id_conflict",
              "sequence_conflict",
              "endpoint_limit",
              "replay_not_eligible",
              "replay_limit",
              "unsupported_media_type",
              "rate_limited",
              "queue_full",
              "quota_exceeded",
              "billing_pending",
              "unavailable"
            ]
          }
        },
        "additionalProperties": false
      },
...
        "SealedMessagePage": {
          "type": "object",
          "additionalProperties": false,
          "required": [
            "messages",
            "next_cursor"
          ],
          "properties": {
            "messages": {
              "type": "array",
              "maxItems": 20,
              "items": {
                "$ref": "#/components/schemas/SealedMessageMetadata"
              }
            },
            "next_cursor": {
              "type": [
                "string",
                "null"
              ],
              "format": "uuid"
            }
          }
        }
    }
  }
}